Add car related features.

+++ app/daos/user.py
@@ -4,6 +4,7 @@ from werkzeug.exceptions import abort
 from app.daos.base import BaseDAO
 from app.database.models.user import UserModel
 from app.helpers.localization import gettext
+from app.daos.demand import demand_dao
 
 
 class UserDAO(BaseDAO):
@@ -25,8 +26,13 @@ class UserDAO(BaseDAO):
         return user
 
     def delete_user(self, user):
-        self.session.delete(user)
-        self.session.commit()
+        has_demands = demand_dao.has_any_demand(user.id)
+
+        if has_demands:
+            Exception(400, {"message": gettext("user_has_demands")})
+        else:
+            self.session.delete(user)
+            self.session.commit()
 
     def check_email_available(self, _email):
         return self.session.query(self.model).filter_by(email=_email).count() == 0
